chg: dev: wip: bootstrap > init > pgpb > pgpb-org > pgpb-org-*
Reviewing end-to-end the way Org mode is configured and restructuring
the code into packages and functions, so that it's more modular and
easier to understand.

The entrypoint for my package is pgpb.el, generated from it's .org
counterpart.

The commit subject indicates the sequence that I hope to achieve, that
pgpb-org.el be the entrypoint for all my Org mode customisations.

Other packages follow the same convention: =pgpb-<topic>=